+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Changed type of stored layer names to `vk::ExtensionName` instead of `CString`
 - Added constants, constructor, and additional conversions to `Version`
 - Changed `patch` field to default to `0` in `Version` instead of using `Option`
+- Reworked byte and string array types in Vulkan structs
 
 ## [0.10.0] - 2021-07-10